Add cli option to specify src directory
